Implement resampling, but commented out likelihood
Since our compute_likelihood function is buggy, I commented out the line
that calls compute_likelihood and proceeded with the initial
probabilities (uniform) so I can try out resample.
I also included a script to compile for Valgrind memtest. The memory
leaks don't seem to come from the code we written. It comes from the
rest of the code we didn't touch and probably the third party libraries.