Fix normalize bug and update compute likelihood
One of the normalize loops forgot to move to the next particle, so only
the first particle in the list was being normalized. It divided its
probability by the total likelihood 10000 times, which either made it 0
or inf. So this explains why incorporating the previous belief in the
compute_likelihood function behaved unexpectedly.
After fixing that bug, we can actually multiply the previous belief by
the probability for the sonar readings as it should be in the algoirthm.
The probability for the sonar readings is now the product of the
relative Gaussian probability of each error. Since the max value of the
Gaussian pdf is around 0.02, we divided by that max so the multipiers
aren't so small.